refactor(cli): the presenter modules live under src/cli/
cli-models.ts, cli-auth.ts, cli-add-feishu.ts and invoke-stream.ts
predate the commander rebuild (#243) and stayed at the package top
level next to the public subpath entries, though every consumer is
CLI-side. Moved under src/cli/ (the cli- prefix is redundant there:
models-view / auth-view / add-feishu / invoke-stream), tests renamed
to match, AGENTS.md map updated. Pure move — no behavior change; the
package-boundary neutrality test tracks invoke-stream at its new
path.
