+++ src/cli/shared.ts
@@ -7,7 +7,7 @@ import { readFile, writeFile } from "node:fs/promises";
 import { relative } from "node:path";
 import { autocomplete, isCancel, log as clackLog, password, select, text as clackText } from "@clack/prompts";
 import type { Models } from "@earendil-works/pi-ai";
-import { buildModelPickerOptions } from "../cli-models.ts";
+import { buildModelPickerOptions } from "./models-view.ts";
 import { fastagentCredentialStore } from "../engines/pi/auth.ts";
 import {
   isValidPort,
@@ -23,7 +23,7 @@ import {
 import { ensureStateRootSelfIgnored, isUnderDir } from "../engines/pi/definition.ts";
 import { LoginCancelled, type LoginIO, type LoginMethod, type LoginResult, loginFlow } from "../engines/pi/login.ts";
 import { createPiModels, probeApiKey, probeAuthSource, providerAuthStatuses } from "../engines/pi/models.ts";
-import { formatAuthReport } from "../cli-auth.ts";
+import { formatAuthReport } from "./auth-view.ts";
 import { log } from "../log.ts";
 import { openExternalUrl } from "../open-url.ts";
 import { failStartup, failUsage } from "./fail.ts";
